+++ CHANGELOG.md
@@ -4,6 +4,19 @@
 
 ***
 
+### [v3.3.0]
+
+**Optimized**:
+  - Optimize openocd flash command-line.
+  - Allow show disasm for `elf`/`axf`.
+
+**Fixed**:
+  - Cannot provide C/C++ intellisense configuration for files that contain symbol links for `linux` platform.
+  - The `exclude source list` in the old project template is invalid for `linux` platform.
+  - Can't parse toolchain system includes and defines for `linux` platform.
+
+***
+
 ### [v3.2.0]
 
 **New**: